Refactor and centralize login service and validation logic.
Replaced `LoginEntryData` with `User` and streamlined login-related services into a dedicated `login` package. Introduced `DefaultPasswordValidator` and improved abstraction for better modularity and maintainability. Updated all related modules for consistency.
