+++ login/default_password_validation.go
@@ -1,0 +1,19 @@
+package login
+
+import (
+	apicontext "github.com/softwareplace/http-utils/context"
+	"github.com/softwareplace/http-utils/security/encryptor"
+)
+
+// DefaultPasswordValidator is a generic type responsible for validating user passwords
+// against their stored encrypted counterparts in principal contexts.
+//
+// T represents a type that implements the Principal interface.
+// It ensures that the principal context contains methods to retrieve the encrypted password and other details.
+type DefaultPasswordValidator[T apicontext.Principal] struct {
+}
+
+func (a *DefaultPasswordValidator[T]) IsValidPassword(loginEntryData User, principal T) bool {
+	return encryptor.NewEncrypt(loginEntryData.Password).
+		IsValidPassword(principal.EncryptedPassword())
+}
