--- server/auth.go
@@ -1,122 +1,0 @@
-package server
-
-import (
-	log "github.com/sirupsen/logrus"
-	apicontext "github.com/softwareplace/http-utils/context"
-	errorhandler "github.com/softwareplace/http-utils/error"
-	"github.com/softwareplace/http-utils/security"
-	"github.com/softwareplace/http-utils/security/encryptor"
-	"time"
-)
-
-type LoginEntryData struct {
-	Username string `json:"username"`
-	Password string `json:"password"`
-	Email    string `json:"email"`
-}
-
-type LoginService[T apicontext.Principal] interface {
-	// SecurityService returns an instance of SecurityService responsible for providing security-related operations.
-	// This includes handling encryption, decryption, JWT generation, claim extraction, and authorization processes.
-	// It ensures the necessary security mechanisms are available in the context of the LoginService.
-	//
-	// Returns:
-	//   - security.Service[T]: The security service instance associated with the implementing service.
-	SecurityService() security.Service[T]
-
-	// Login processes the login request for the specified user by validating their credentials.
-	// It authenticates the user based on the provided login data and returns an authenticated principal context or an error.
-	//
-	// Parameters:
-	//   - user: An instance of LoginEntryData that contains the username, encryptor, and/or email for user authentication.
-	//
-	// Returns:
-	//   - T: The authenticated principal context representing the logged-in user.
-	//   - error: If authentication fails, an error is returned.
-	Login(user LoginEntryData) (T, error)
-
-	// TokenDuration specifies the duration for which a generated JWT token remains valid.
-	// This value defines the time-to-live (TTL) for the token, ensuring secure and proper session management.
-	//
-	// Returns:
-	//   - time.Duration: The duration for which a generated token is valid.
-	TokenDuration() time.Duration
-
-	// IsValidPassword validates the user-provided plaintext password against the stored encrypted password.
-	//
-	// This method uses the encryptor package to create a password hash from the provided loginEntryData password
-	// and compares it with the encrypted password available in the principal context.
-	//
-	// A default implementation is available as server.DefaultPasswordValidator[T]
-	//
-	// Parameters:
-	//   - loginEntryData: The LoginEntryData containing the plaintext password to be validated.
-	//   - principal: The principal context of type T, which contains the stored encrypted password.
-	//
-	// Returns:
-	//   - bool: True if the passwords match; false otherwise.
-	IsValidPassword(loginEntryData LoginEntryData, principal T) bool
-}
-
-func (a *baseServer[T]) Login(ctx *apicontext.Request[T]) {
-	GetRequestBody(ctx, LoginEntryData{}, a.loginDataHandler, FailedToLoadBody[T])
-}
-
-func (a *baseServer[T]) ApiKeyGenerator(ctx *apicontext.Request[T]) {
-	GetRequestBody(ctx, ApiKeyEntryData{}, a.apiKeyGeneratorDataHandler, FailedToLoadBody[T])
-}
-
-func (a *baseServer[T]) loginDataHandler(ctx *apicontext.Request[T], loginEntryData LoginEntryData) {
-
-	errorhandler.Handler(func() {
-		loginService := a.loginService
-		decrypt, err := loginService.SecurityService().Decrypt(loginEntryData.Password)
-		if err != nil {
-			log.Printf("LOGIN/DECRYPT: Failed to decrypt encryptor: %v", err)
-		} else {
-			loginEntryData.Password = decrypt
-		}
-
-		principal, err := loginService.Login(loginEntryData)
-
-		if err != nil {
-			log.Printf("LOGIN/LOGIN: Failed to login: %v", err)
-			ctx.Forbidden("Login failed: Invalid username or password")
-			return
-		}
-
-		isValidPassword := loginService.IsValidPassword(loginEntryData, principal)
-
-		if !isValidPassword {
-			log.Printf("LOGIN/LOGIN: Failed to login: %v", err)
-			ctx.Forbidden("Login failed: Invalid username or password")
-			return
-		}
-
-		jwt, err := loginService.SecurityService().GenerateJWT(principal, loginService.TokenDuration())
-
-		if err != nil {
-			log.Printf("LOGIN/JWT: Failed to generate JWT: %v", err)
-			ctx.InternalServerError("Login failed with internal server error. Please try again later.")
-			return
-		}
-
-		ctx.Ok(jwt)
-	}, func(err error) {
-		log.Printf("LOGIN/HANDLER: Failed to handle request: %v", err)
-		ctx.Forbidden("Login failed: Invalid username or password")
-	})
-}
-
-// DefaultPasswordValidator is a generic type responsible for validating user passwords
-// against their stored encrypted counterparts in principal contexts.
-//
-// T represents a type that implements the Principal interface.
-// It ensures that the principal context contains methods to retrieve the encrypted password and other details.
-type DefaultPasswordValidator[T apicontext.Principal] struct {
-}
-
-func (a *DefaultPasswordValidator[T]) IsValidPassword(loginEntryData LoginEntryData, principal T) bool {
-	return encryptor.NewEncrypt(loginEntryData.Password).
-		IsValidPassword(principal.EncryptedPassword())
-}
